--- lib/pages/page_1.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-import '../presentation/general/card_item.dart';
-
-class PageWidget1 extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: <Widget>[
-        Row(
-          children: [CardItem(Colors.teal, text: 'Changed teal',),],
-        ),
-      ],
-    );
-  }
-}
