+++ lib/presentation/general/general_page_view.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:movies/di/injector.dart';
-import '../../pages/page_0.dart';
-import '../../pages/page_1.dart';
+import '../../screens/popular_movies.dart';
+import '../../screens/top_rated_movies.dart';
+import '../general/card/card_item.dart';
 
 class GeneralPageView extends StatefulWidget {
   @override
@@ -10,8 +11,17 @@ class GeneralPageView extends StatefulWidget {
 
 class _GeneralPageViewState extends State<GeneralPageView> {
   PageController _controller = PageController(
-    initialPage: 1,
+    initialPage: 0,
   );
+
+  final List<Widget> testData1 = [
+    CardItem(Colors.teal, text: 'Changed teal'),
+    Container(
+      padding: const EdgeInsets.all(8),
+      child: const Text('Heed not the rabble'),
+      color: Colors.green[200],
+    ),
+  ];
 
   @override
   Widget build(BuildContext context) {
@@ -19,8 +29,8 @@ class _GeneralPageViewState extends State<GeneralPageView> {
     return PageView(
       controller: _controller,
       children: [
-        PageWidget0(),
-        PageWidget1(),
+        PopularMoviesWidget(testData1),
+        TopRatedMoviesWidget([testData1[1], testData1[1], testData1[1]]),
       ],
     );
   }
