+++ lib/presentation/main.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:movies/di/injector.dart';
 import 'general/general_page_view.dart';
+import 'general/app_bar/app_bar.dart';
 
 void main() {
-  var injector = Injector(child : MoviesApp());
+  var injector = Injector(child: MoviesApp());
   injector.init();
   runApp(injector);
 }
 
 class MoviesApp extends StatelessWidget {
-
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -17,25 +17,9 @@ class MoviesApp extends StatelessWidget {
           primarySwatch: Colors.deepPurple, brightness: Brightness.dark),
       debugShowCheckedModeBanner: false,
       home: Scaffold(
-        appBar: AppBar(
-          backgroundColor: Colors.deepPurple,
-          actions: <Widget>[
-            IconButton(
-              icon: const Icon(Icons.menu),
-              tooltip: 'Menu',
-              onPressed: () {},
-            ),
-            IconButton(
-              icon: const Icon(Icons.more_vert),
-              tooltip: 'More',
-              onPressed: () {},
-            ),
-          ],
-          title: const Text('AppBar'),
-        ),
+        appBar: AppBarWidget(),
         body: GeneralPageView(),
       ),
     );
   }
-
 }
