+++ lib/screens/top_rated_movies.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+import '../presentation/general/scroll_view/scroll_view.dart';
+
+class TopRatedMoviesWidget extends StatelessWidget {
+  final List<Widget> items;
+  TopRatedMoviesWidget(this.items);
+
+  @override
+  Widget build(BuildContext context) {
+    return ScrolViewWidget(items);
+  }
+}
